// web/src/shared/components/layout/Button/types.ts
export enum ButtonSize {
  LARGE = 'large',
  SMALL = 'small',
}

export enum ButtonStyleVariant {
  // Add in figma
  PRIMARY = 'variant-primary',
  CONFIRM = 'variant-confirm',
  // Create in figma
  LINK = 'variant-link',
  SAVE = 'variant-save',
  DELETE = 'variant-delete',
  ICON = 'variant-icon',
  // default, link variant without hover changes
  STANDARD = 'variant-standard',
}